refactor(viewer): extract useViewerUrlState from GeoViewer
Decomposition 3/4, content-neutral: the one-shot layer-stack restore,
the debounced live report, and the camera listener move into an
effects-only hook taking the mount snapshot as an explicit "initial"
param. Camera restore stays with View creation and time seeds with
useViewerTimeline. Tests untouched.
