+++ frontend/src/features/viewer/geo/GeoViewer.tsx
@@ -33,8 +33,7 @@ import {
 import { useTranslation } from 'react-i18next'
 import 'ol/ol.css'
 import { RefreshCw } from 'lucide-react'
-import { fromLonLat, toLonLat } from 'ol/proj'
-import { unByKey } from 'ol/Observable'
+import { fromLonLat } from 'ol/proj'
 import { useLensSource } from '../hooks/useLensSource'
 import { AUTOFIT_KEY, createViewerView } from '../hooks/useOlMapBase'
 import { formatStep } from '../format'
@@ -56,6 +55,7 @@ import { GeoExportDialog } from './GeoExportDialog'
 import { CompareHelpDialog } from './CompareHelpDialog'
 import { AnnotationEditorDialog } from './AnnotationEditorDialog'
 import { useViewerAnnotations } from './useViewerAnnotations'
+import { useViewerUrlState } from './useViewerUrlState'
 import { useViewerTimeline } from './useViewerTimeline'
 import { downloadAnnotationsGeojson } from './annotations'
 import { useGeoShortcuts } from './useGeoShortcuts'
@@ -179,84 +179,6 @@ export function GeoViewer({
 
   const activeOrderA = selection.activeOrderFor('a')
   const activeOrderB = selection.activeOrderFor('b')
-
-  // -------- URL view-state restore (one-shot per slot) --------
-  const pendingLayersRef = useRef<{
-    a: ReadonlyArray<string>
-    b: ReadonlyArray<string>
-    unlinked: boolean
-  } | null>(
-    initialViewRef.current?.layersA?.length ||
-      initialViewRef.current?.layersB?.length
-      ? {
-          a: initialViewRef.current.layersA ?? [],
-          b: initialViewRef.current.layersB ?? [],
-          unlinked: initialViewRef.current.unlinkedLayers === true,
-        }
-      : null,
-  )
-  // Mirrors the ref as state so the report effect re-runs on completion.
-  const [restorePending, setRestorePending] = useState({
-    a: (pendingLayersRef.current?.a.length ?? 0) > 0,
-    b: (pendingLayersRef.current?.b.length ?? 0) > 0,
-  })
-  useEffect(() => {
-    const pending = pendingLayersRef.current
-    if (!pending) return
-    // Switch the selection model first — toggles must land per-side.
-    if (pending.unlinked && selection.linkMode === 'linked') {
-      selection.setLinkMode('unlinked')
-      return
-    }
-    // Both slots can settle in one pass — stale isPairActive would re-toggle.
-    const toggledNow = new Set<string>()
-    for (const [slot, source] of [
-      ['a', sourceA],
-      ['b', sourceB],
-    ] as const) {
-      const names = pending[slot]
-      if (names.length === 0 || source.loadingLayers) continue
-      // B may still be starting; if it never runs the URL keeps the value.
-      if (slot === 'b' && !hasB) continue
-      const available = new Set(source.layers.map((l) => l.name))
-      // Reverse: toggles prepend, so the first name ends up on top.
-      for (const name of [...names].reverse()) {
-        if (!available.has(name)) continue
-        if (pending.unlinked) {
-          if (!selection.isLayerActive(slot, name)) {
-            selection.toggleLayer(slot, name)
-          }
-        } else {
-          const pair = pairing.pairs.find(
-            (p) => p.perSource[slot]?.name === name,
-          )
-          if (
-            pair &&
-            !toggledNow.has(pair.key) &&
-            !selection.isPairActive(pair.key)
-          ) {
-            toggledNow.add(pair.key)
-            selection.togglePair(pair.key)
-          }
-        }
-      }
-      pending[slot] = []
-    }
-    if (pending.a.length === 0 && pending.b.length === 0) {
-      pendingLayersRef.current = null
-    }
-    setRestorePending((prev) => {
-      const next = { a: pending.a.length > 0, b: pending.b.length > 0 }
-      return prev.a === next.a && prev.b === next.b ? prev : next
-    })
-    // Meaningful bits only — selection/source identities churn every render.
-  }, [
-    sourceA.loadingLayers,
-    sourceB.loadingLayers,
-    hasB,
-    pairing.pairs,
-    selection.linkMode,
-  ])
 
   // Opacity hierarchy: global × per-source × per-layer (per-layer lives in
   // the selection; the product of the first two feeds the map stacks).
@@ -475,53 +397,23 @@ export function GeoViewer({
     }
   }, [availableBasemaps, basemapId, sourceA.loadingLayers])
 
-  // -------- URL view-state report (page debounces into the URL) --------
-  useEffect(() => {
-    if (!onViewStateChange) return
-    const partial: Partial<ViewerUrlState> = {
-      // Auto-unlink is situational — persisting it would block the relink.
-      unlinkedLayers:
-        selection.linkMode === 'unlinked' && !selection.autoUnlinked,
-      timeLink: timeLinkMode,
-      offsetMs,
-      basemap: basemapId === DEFAULT_BASEMAP_ID ? undefined : basemapId,
-    }
-    // Hold restored fields until slots settle — mid-load writes would strip them.
-    if (!restorePending.a) partial.layersA = activeOrderA
-    if (!restorePending.b) partial.layersB = activeOrderB
-    if (!restorePending.a && !restorePending.b) {
-      partial.timeMs = currentEpoch ?? undefined
-    }
-    onViewStateChange(partial)
-  }, [
+  // -------- URL view-state restore + report --------
+  useViewerUrlState({
+    initial: initialViewRef.current,
     onViewStateChange,
+    viewRef,
+    selection,
+    pairing,
+    sourceA,
+    sourceB,
+    hasB,
     activeOrderA,
     activeOrderB,
-    selection.linkMode,
-    selection.autoUnlinked,
-    restorePending,
     currentEpoch,
     timeLinkMode,
     offsetMs,
     basemapId,
-  ])
-  useEffect(() => {
-    const view = viewRef.current
-    if (!onViewStateChange || !view) return
-    const report = () => {
-      const center = view.getCenter()
-      const zoom = view.getZoom()
-      if (!center || zoom === undefined) return
-      const [lon, lat] = toLonLat(center)
-      if (![lon, lat, zoom].every(Number.isFinite)) return
-      onViewStateChange({ camera: { lon, lat, zoom } })
-    }
-    const keys = [
-      view.on('change:center', report),
-      view.on('change:resolution', report),
-    ]
-    return () => unByKey(keys)
-  }, [onViewStateChange])
+  })
 
   // Time-step prefetch (default off — bandwidth-heavy).
   const [preloadTimeSteps, setPreloadTimeSteps] = useState(false)
